requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26